Fixes for CC1101

+++ components/wmbus_common/meters.cpp
@@ -474,7 +474,7 @@ std::string MeterCommonImplementation::datetimeOfUpdateRobot() {
 std::string MeterCommonImplementation::unixTimestampOfUpdate() {
   char ut[40];
   memset(ut, 0, sizeof(ut));
-  snprintf(ut, sizeof(ut) - 1, "%lu", datetime_of_update_);
+  snprintf(ut, sizeof(ut) - 1, "%llu", (unsigned long long)datetime_of_update_);
   return std::string(ut);
 }
